requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
openai-clip
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.6
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
ftfy==6.3.1
grpcio==1.69.0
h11==0.14.0
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.2.1
numpy==2.2.1
openai-clip==1.0.1
packaging==24.2
pillow==11.0.0
protobuf==5.29.3
psycopg2==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.1
setuptools==70.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch
torchvision
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
